+++ R/ConditionalSmoothedScatterPlot.R
@@ -37,7 +37,9 @@ smoothing = function(frm, xvar, yvar, k, align) {
 }
 
 
-#' Plot a scatter plot with smoothing line, with smoothing window aligned either left, center or right,
+#' Plot a scatter plot with smoothing line.
+#'
+#' Plot a scatter plot with smoothing line with smoothing window aligned either left, center or right,
 #'       xvar is the continuous independent variable and yvar is the dependent binary variable.
 #'       Smoothing is by a square window of width k
 #'
@@ -53,15 +55,24 @@ smoothing = function(frm, xvar, yvar, k, align) {
 #'
 #' y = c(1,2,3,4,5,10,15,18,20,25)
 #' x = seq_len(length(y))
-#' df = data.frame(x=x,y=y)
+#' df = data.frame(x=x, y=y, group=x>5)
 #' WVPlots::ConditionalSmoothedScatterPlot(df, "x", "y", NULL,
-#'    title="left smooth, one group", align="left")
+#'    title="left smooth, one groups", align="left")
+#' WVPlots::ConditionalSmoothedScatterPlot(df, "x", "y", "group",
+#'    title="left smooth, two groups", align="left")
 #'
 #' @export
 ConditionalSmoothedScatterPlot = function(frame, xvar, yvar, groupvar, title, ...,
                                           k=3, align="center") {
-  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ConditionalSmoothedScatterPlot")
-  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title)
+  vlist <- list(xvar = xvar, yvar = yvar)
+  if(!is.null(groupvar)) {
+    vlist$groupvar <- groupvar
+  }
+  frame <- check_frame_args_list(...,
+                               frame = frame,
+                               name_var_list = vlist,
+                               title = title,
+                               funname = "WVPlots::ConditionalSmoothedScatterPlot")
   if(!is.null(groupvar)) {
     if(!isDiscrete(frame[[groupvar]])) {
       stop(paste(groupvar, "should be discrete (factor, character, integer, or logical)"))