+++ R/PRPlot.R
@@ -70,8 +70,11 @@ calcPR <- function(modelPredictions,yValues) {
 #'
 #' @export
 PRPlot <- function(frame, xvar, truthVar, truthTarget, title,...) {
-  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::PRPlot")
-  checkArgs(frame=frame,xvar=xvar,yvar=truthVar,title=title)
+  frame <- check_frame_args_list(...,
+                                 frame = frame,
+                                 name_var_list = list(xvar = xvar, truthVar = truthVar),
+                                 title = title,
+                                 funname = "WVPlots::PRPlot")
   outcol <- frame[[truthVar]]==truthTarget
   if(length(unique(outcol))!=2) {
     return(NULL)