+++ R/ScatterHistC.R
@@ -30,8 +30,11 @@ ScatterHistC = function(frame, xvar, yvar, cvar, title, ...,
      (!requireNamespace("gridExtra", quietly = TRUE))) {
     return("WVPlots::ScatterHistC requires the grid and gridExtra packages be installed")
   }
-  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ScatterHistC")
-  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title)
+  frame <- check_frame_args_list(...,
+                                 frame = frame,
+                                 name_var_list = list(xvar = xvar, yvar = yvar, cvar = cvar),
+                                 title = title,
+                                 funname = "WVPlots::ScatterHistC")
   minimal_labels = TRUE
 
   # Use this plot to print the legend.
@@ -197,8 +200,11 @@ ScatterHistN = function(frame, xvar, yvar, zvar, title, ...,
      (!requireNamespace("gridExtra", quietly = TRUE))) {
     return("WVPlots::ScatterHistN requires the grid and gridExtra packages be installed")
   }
-  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ScatterHistN")
-  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title)
+  frame <- check_frame_args_list(...,
+                                 frame = frame,
+                                 name_var_list = list(xvar = xvar, yvar = yvar, zvar = zvar),
+                                 title = title,
+                                 funname = "WVPlots::ScatterHistN")
   q <- sort(unique(quantile(frame[[zvar]],seq(0, 1, 1/nclus))))
   yC <- cut(frame[[zvar]],q,include.lowest=TRUE)
   if(length(unique(yC))<=1) {