+++ R/ShadowHist.R
@@ -34,8 +34,8 @@
 ShadowHist = function(frm, xvar, condvar, title, ...,
                       ncol = 1, palette = "Dark2",
                       bins = 30, binwidth = NULL) {
-  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ShadowHist")
-  frm <- check_frame_args_list(frame = frm,
+  frm <- check_frame_args_list(...,
+                               frame = frm,
                                name_var_list = list(xvar = xvar, condvar = condvar),
                                title = title,
                                funname = "WVPlots::ShadowHist")