+++ NEWS.md
@@ -1,9 +1,10 @@
 
-# WVPlots 0.3.3 2018-05-18
+# WVPlots 0.3.3 2018-05-21
 
  * Remove F-test from ScatterHist.
  * Added shadow plot.
  * Improve argument checking.
+ * Add frame narrowing.
  
 # WVPlots 0.3.2 2018-05-07